@@ -1,11 +1,10 @@
 import flet as ft
 from loguru import logger
-from sqlmodel import SQLModel
 
-from config import APP_TITLE, engine
+from config import APP_TITLE, create_db_and_tables
 from env import setup_environment
 from logging_conf import setup_logger
-from router import Router
+from router_config import setup_enhanced_routing
 
 # 環境変数ファイルの作成
 setup_environment()
@@ -15,7 +14,7 @@ setup_logger()
 logger.info("アプリケーションを起動します。")
 
 # DB初期化（全テーブル作成）
-SQLModel.metadata.create_all(engine)
+create_db_and_tables()
 
 
 def main(page: ft.Page) -> None:
@@ -26,7 +25,17 @@ def main(page: ft.Page) -> None:
     """
     # ページの初期設定
     page.title = APP_TITLE
-    Router(page)
+    page.theme_mode = ft.ThemeMode.LIGHT
+    page.padding = 0
+    page.theme = ft.Theme(
+        color_scheme_seed=ft.Colors.GREY_700,
+        page_transitions=ft.PageTransitionsTheme(
+            windows=ft.PageTransitionTheme.NONE,
+        ),
+    )
+
+    # FletNativeRouterを使用したルーティング設定
+    setup_enhanced_routing(page)
 
     logger.info("セッションが開始されました。")
 
