@@ -1,11 +1,22 @@
-# model/task.py
-# SQLModelとSQLiteを使ったタスク管理モデル
-from __future__ import annotations
+"""新しいTaskモデルの定義"""
 
-from datetime import datetime
+import uuid
+from datetime import date
+from enum import Enum
 
 from sqlmodel import Field, SQLModel
-from typing_extensions import deprecated
+
+
+class TaskStatus(str, Enum):
+    """タスクのステータス（GTDシステムに基づく）"""
+
+    INBOX = "inbox"
+    NEXT_ACTION = "next_action"
+    WAITING_FOR = "waiting_for"
+    SOMEDAY_MAYBE = "someday_maybe"
+    DELEGATED = "delegated"
+    COMPLETED = "completed"
+    CANCELLED = "cancelled"
 
 
 class TaskBase(SQLModel):
@@ -14,102 +25,90 @@ class TaskBase(SQLModel):
     タスクの基本情報を定義するモデルクラス。SQLModelを使用してデータベースと連携します。
 
     Attributes:
+        project_id (uuid.UUID | None): タスクが属するプロジェクトのID。プロジェクトに属さないタスクの場合はNone。
+        parent_id (uuid.UUID | None): 親タスクのID。サブタスクの場合に使用。
         title (str): タスクのタイトル。インデックスが設定されており、検索に使用されます。
         description (str): タスクの詳細説明。デフォルトは空文字列。
-        created_at (datetime): タスクの作成日時。デフォルトは現在の日時。
-        updated_at (datetime): タスクの更新日時。デフォルトは現在の日時。
-        completed (bool): タスクの完了状態。デフォルトはFalse（未完了）。
+        status (TaskStatus): タスクのステータス。デフォルトはINBOX。
+        due_date (date | None): タスクの締め切り日。設定されていない場合はNone。
     """
 
+    project_id: uuid.UUID | None = Field(default=None, foreign_key="project.id", index=True)
+    parent_id: uuid.UUID | None = Field(default=None, foreign_key="task.id", index=True)
     title: str = Field(index=True)
     description: str = Field(default="")
-    created_at: datetime = Field(default_factory=datetime.now)
-    updated_at: datetime = Field(default_factory=datetime.now)
-    completed: bool = Field(default=False)
+    status: TaskStatus = Field(default=TaskStatus.INBOX, index=True)
+    due_date: date | None = Field(default=None, index=True)
 
 
 class Task(TaskBase, table=True):
-    """タスクモデル
+    """新しいタスクモデル
 
     タスクの情報をデータベースに保存するためのモデルクラス。SQLModelを使用してデータベースと連携します。
 
     Attributes:
-        id (int | None): タスクのID。デフォルトはNoneで、データベースに保存時に自動生成されます。
+        id (uuid.UUID | None): タスクのID。デフォルトはNoneで、データベースに保存時に自動生成されます。
+        project_id (uuid.UUID | None): タスクが属するプロジェクトのID。プロジェクトに属さないタスクの場合はNone。
+        parent_id (uuid.UUID | None): 親タスクのID。サブタスクの場合に使用。
         title (str): タスクのタイトル。インデックスが設定されており、検索に使用されます。
         description (str): タスクの詳細説明。デフォルトは空文字列。
-        created_at (datetime): タスクの作成日時。デフォルトは現在の日時。
-        updated_at (datetime): タスクの更新日時。デフォルトは現在の日時。
-        completed (bool): タスクの完了状態。デフォルトはFalse（未完了）。
+        status (TaskStatus): タスクのステータス。デフォルトはINBOX。
+        due_date (date | None): タスクの締め切り日。設定されていない場合はNone。
     """
 
-    id: int | None = Field(default=None, primary_key=True)
+    id: uuid.UUID | None = Field(default_factory=uuid.uuid4, primary_key=True)
 
 
 class TaskCreate(TaskBase):
-    """タスク作成用モデル
+    """新しいタスク作成用モデル
 
     タスクを新規作成する際に使用するモデルクラス。SQLModelを使用してデータベースと連携します。
 
     Attributes:
+        project_id (uuid.UUID | None): タスクが属するプロジェクトのID。プロジェクトに属さないタスクの場合はNone。
+        parent_id (uuid.UUID | None): 親タスクのID。サブタスクの場合に使用。
         title (str): タスクのタイトル。インデックスが設定されており、検索に使用されます。
         description (str): タスクの詳細説明。デフォルトは空文字列。
-        created_at (datetime): タスクの作成日時。デフォルトは現在の日時。
-        updated_at (datetime): タスクの更新日時。デフォルトは現在の日時。
-        completed (bool): タスクの完了状態。デフォルトはFalse（未完了）。
+        status (TaskStatus): タスクのステータス。デフォルトはINBOX。
+        due_date (date | None): タスクの締め切り日。設定されていない場合はNone。
     """
 
 
 class TaskRead(TaskBase):
-    """タスク読み取り用モデル
+    """新しいタスク読み取り用モデル
 
     タスクの情報を読み取る際に使用するモデルクラス。SQLModelを使用してデータベースと連携します。
 
     Attributes:
-        id (int): タスクのID。
+        id (uuid.UUID): タスクのID。
+        project_id (uuid.UUID | None): タスクが属するプロジェクトのID。プロジェクトに属さないタスクの場合はNone。
+        parent_id (uuid.UUID | None): 親タスクのID。サブタスクの場合に使用。
         title (str): タスクのタイトル。インデックスが設定されており、検索に使用されます。
         description (str): タスクの詳細説明。デフォルトは空文字列。
-        created_at (datetime): タスクの作成日時。デフォルトは現在の日時。
-        updated_at (datetime): タスクの更新日時。デフォルトは現在の日時。
-        completed (bool): タスクの完了状態。デフォルトはFalse（未完了）。
+        status (TaskStatus): タスクのステータス。デフォルトはINBOX。
+        due_date (date | None): タスクの締め切り日。設定されていない場合はNone。
     """
 
-    id: int
+    id: uuid.UUID
 
 
 class TaskUpdate(SQLModel):
-    """タスク更新用モデル
+    """新しいタスク更新用モデル
 
     タスクの情報を更新する際に使用するモデルクラス。SQLModelを使用してデータベースと連携します。
 
     Attributes:
+        project_id (uuid.UUID | None): タスクが属するプロジェクトのID。Noneの場合は更新しない。
+        parent_id (uuid.UUID | None): 親タスクのID。Noneの場合は更新しない。
         title (str | None): タスクのタイトル。Noneの場合は更新しない。
         description (str | None): タスクの詳細説明。Noneの場合は更新しない。
-        completed (bool | None): タスクの完了状態。Noneの場合は更新しない。
+        status (TaskStatus | None): タスクのステータス。Noneの場合は更新しない。
+        due_date (date | None): タスクの締め切り日。Noneの場合は更新しない。
     """
 
+    project_id: uuid.UUID | None = None
+    parent_id: uuid.UUID | None = None
     title: str | None = None
     description: str | None = None
-    completed: bool | None = None
-
-
-@deprecated(
-    "validate_task_idは非推奨です。TaskRead.idを直接使用してください。",
-)
-def validate_task_id(task_id: int | None) -> int:
-    """タスクIDのチェック
-
-    task.idがPylanceの警告を回避するための関数
-
-    Args:
-        task_id: チェックするタスクのID
-
-    Returns:
-        int: 有効なタスクID
-
-    Raises:
-        ValueError: タスクIDがNoneまたは0の場合
-    """
-    if task_id is None or task_id <= 0:
-        e_msg = "タスクIDが指定されていません、または無効です"
-        raise ValueError(e_msg)
-    return task_id
+    status: TaskStatus | None = None
+    due_date: date | None = None
