@@ -3,14 +3,15 @@ import flet as ft
 from config import APP_TITLE
 
 
-def app_bar(page: ft.Page) -> ft.AppBar:
+def app_bar(page: ft.Page, page_name: str | None = None) -> ft.AppBar:
     """アプリケーションのヘッダーを生成"""
+    title = f"{APP_TITLE} - {page_name}" if page_name else APP_TITLE
     return ft.AppBar(
-        leading=ft.IconButton(ft.icons.HOME, on_click=lambda _: page.go("/")),
-        title=ft.Text(APP_TITLE),
+        leading=ft.IconButton(ft.Icons.HOME, on_click=lambda _: page.go("/")),
+        title=ft.Text(title),
         center_title=False,
         actions=[
-            ft.IconButton(ft.icons.HOME, on_click=lambda _: page.go("/")),
+            ft.IconButton(ft.Icons.HOME, on_click=lambda _: page.go("/")),
             ft.PopupMenuButton(
                 items=[
                     ft.PopupMenuItem(
