@@ -1,1 +1,10 @@
-"""ロジックテスト用パッケージ."""
+"""Logic層テストパッケージ
+
+このパッケージは、logic層の最重要部分（Repository層とService層）に対する
+テストコードを提供します。
+
+最低限版として以下の要素をテスト対象とします：
+- BaseRepository: 共通CRUD操作
+- TaskRepository: タスク固有のデータアクセス
+- TaskService: タスクのビジネスロジック
+"""
